Convert portfolio page to page model.  Refactor page controller to service
